toast style, move i18n
