+++ services/tenant-ui/frontend/src/assets/style.scss
@@ -3,3 +3,4 @@
 @import 'forms.scss';
 @import 'tenantuiComponents.scss';
 @import 'primevueComponents.scss';
+@import 'toast.scss';
