+++ services/tenant-ui/frontend/src/plugins/i18n/i18n.ts
