+++ services/tenant-ui/frontend/src/plugins/i18n/locales/jp.json
